+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/Common.kt
@@ -415,6 +415,7 @@ const val mediaField = "media"
 const val disableEditMessageField = "disable_edit_message"
 const val scoreField = "score"
 const val forceField = "force"
+const val forceReplyField = "force_reply"
 
 const val regularPollType = "regular"
 const val quizPollType = "quiz"